refactor(debug): move the exit-related items to a dedicated module
