+++ src/ariel-os-debug/src/exit.rs
@@ -1,0 +1,51 @@
+/// Represents the exit code of a debug session.
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum ExitCode {
+    #[doc(hidden)]
+    Success,
+    #[doc(hidden)]
+    Failure,
+}
+
+impl ExitCode {
+    /// The [`ExitCode`] for success.
+    pub const SUCCESS: Self = Self::Success;
+    /// The [`ExitCode`] for failure.
+    pub const FAILURE: Self = Self::Failure;
+
+    #[allow(dead_code, reason = "not always used due to conditional compilation")]
+    fn to_semihosting_code(self) -> i32 {
+        match self {
+            Self::Success => 0,
+            Self::Failure => 1,
+        }
+    }
+    #[allow(dead_code, reason = "not always used due to conditional compilation")]
+    fn to_std_code(self) -> i32 {
+        match self {
+            Self::Success => 0,
+            Self::Failure => 1,
+        }
+    }
+}
+
+/// Terminates the debug output session.
+///
+/// # Note
+///
+/// This may or may not stop the MCU.
+pub fn exit(code: ExitCode) {
+    #[cfg(feature = "semihosting")]
+    semihosting::process::exit(code.to_semihosting_code());
+
+    #[cfg(feature = "std")]
+    std::process::exit(code.to_std_code());
+
+    #[allow(unreachable_code, reason = "stop nagging")]
+    let _ = code;
+
+    #[allow(unreachable_code, reason = "fallback loop")]
+    loop {
+        core::hint::spin_loop();
+    }
+}
